Update README.md
Removing the part about WebdriverIO having a type definition file included. Because when v5 first launched there was only 1 type definition file in webdriverio package. But as of v5.6.0 that type definition has been changed to be async and the sync type definition is in @wdio/sync. I don't want to have to explain that distinction on the README so just removed this line.